requirements.txt
vllm==0.14.0
kafka-python
PyYAML
Pillow
